@@ -17,17 +17,17 @@ dependencies:
   sqflite: ^2.0.2+1
   path: ^1.8.0
 
-  objectbox: ^2.0.0
+  objectbox: ^4.0.1
   objectbox_flutter_libs: any
 
   shared_preferences: ^2.0.6
-  intl: ^0.18.0
+  intl: ^0.19.0
   wakelock_plus: ^1.1.2
 
   url_launcher: ^6.0.6
-  share_plus: ^7.1.0
+  share_plus: ^9.0.0
 
-  youtube_player_flutter: ^8.0.0
+  youtube_player_flutter: ^9.0.1
   just_audio: ^0.9.27
 
   flutter_cached_pdfview: ^0.4.1
@@ -35,7 +35,7 @@ dependencies:
   flutter_svg: ^2.0.6
   flutter_staggered_grid_view: ^0.7.0
 
-  adaptive_dialog: ^1.9.0-no-macos.2
+  adaptive_dialog: ^2.1.0
 
   dotted_border: ^2.0.0+2
 
@@ -45,25 +45,25 @@ dependencies:
 
   cupertino_icons: ^1.0.3
 
-  package_info_plus: ^4.1.0
+  package_info_plus: ^8.0.0
   flutter_slidable: ^3.0.0
-  firebase_core: ^2.4.1
-  firebase_analytics: ^10.1.0
+  firebase_core: ^3.2.0
+  firebase_analytics: ^11.2.0
   cached_network_image: ^3.2.2
-  photo_view: ^0.14.0
+  photo_view: ^0.15.0
   rxdart: ^0.27.7
-  flat_buffers: ^2.0.5
+  flat_buffers: ^23.5.26
   collection: ^1.16.0
   http: ^1.0.0
   freezed_annotation: ^2.2.0
   flutter_riverpod: ^2.3.6
   riverpod_annotation: ^2.1.1
   path_provider: ^2.0.15
-  flutter_quill: ^8.1.6
+  flutter_quill: ^10.0.4
   flex_color_picker: ^3.3.0
   flutter_speed_dial: ^7.0.0
   flutter_native_splash: ^2.3.5
-  sentry_flutter: ^7.12.0
+  sentry_flutter: ^8.2.0
   graphql_flutter: ^5.2.0-beta.6
 
 dev_dependencies:
